src-gpu/requirements.txt
numpy==1.16.3
numpy-stl==2.3.2
opencv-python==4.2.0.32
python-telegram-bot==12.4.2
torch==1.3.1
torchtext==0.5.0
torchvision==0.4.2+cpu
tqdm==4.42.1

